requirements/base.txt
pytz==2023.4  # https://github.com/stub42/pytz
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.6.0  # https://github.com/evansd/whitenoise

# Django
# ------------------------------------------------------------------------------
django==5.0.6  # https://www.djangoproject.com/
django-environ==0.11.2 # https://github.com/joke2k/django-environ
django-model-utils==4.4.0  # https://github.com/jazzband/django-model-utils
requests~=2.31.0
wsgi-basic-auth~=1.1.0
ds-caselaw-marklogic-api-client~=23.0.1
ds-caselaw-utils==1.4.1
rollbar
django-weasyprint==2.3.0
django-waffle==4.1.0  # https://github.com/django-waffle/django-waffle
django-formtools~=2.5.1
crispy-forms-gds~=0.3.1
markdown-it-py~=3.0.0
mdit-py-plugins~=0.4.0

# Type stubs
mypy-boto3-s3==1.34.105
mypy-boto3-sns==1.34.44
